feat(admin-dashboard): added admin dashboard analytics, added also meter_no and meters table
