+++ database/seeders/MeterReadingSeeder.php
@@ -2,6 +2,7 @@
 
 namespace Database\Seeders;
 
+use App\Models\Meter;
 use App\Models\MeterReading;
 use Illuminate\Database\Console\Seeds\WithoutModelEvents;
 use Illuminate\Database\Seeder;
@@ -13,8 +14,13 @@ class MeterReadingSeeder extends Seeder
      */
     public function run(): void
     {
-         MeterReading::create([
-            'customer_id' => 1,          // assumes you have a customer with ID 1
+        $meter = Meter::create([
+            'meter_no' => 'MTR-001',
+            'customer_id' => 1,
+        ]);
+
+        MeterReading::create([
+            'meter_id' => $meter->id,
             'month' => 'January',
             'year' => 2025,
             'meter_value' => 123.45,
